+++ core/src/main/java/org/modelmapper/convention/NameTransformers.java
@@ -63,4 +63,47 @@ public class NameTransformers {
       return "Javabeans Mutator";
     }
   };
+
+  /**
+   * Creates NameTransformer for builder.
+   * @return a NameTransformer
+   */
+  public static NameTransformer builder() {
+    return builder("");
+  }
+
+  /**
+   * Creates NameTransformer for builder.
+   *
+   * @param prefix the prefix for the setter of the builder
+   * @return a NameTransformer
+   */
+  public static NameTransformer builder(String prefix) {
+    return new BuilderNameTransformer(prefix);
+  }
+
+  /**
+   * Transforms the names to their simple property name according to the builder convention.
+   * Class and field names are unchanged.
+   */
+  private static class BuilderNameTransformer implements NameTransformer {
+    private String prefix;
+
+    private BuilderNameTransformer(String prefix) {
+      this.prefix = prefix;
+    }
+
+    public String transform(String name, NameableType nameableType) {
+      if (prefix.isEmpty())
+        return name;
+      if (name.startsWith(prefix))
+        return Strings.decapitalize(name.substring(prefix.length()));
+      return name;
+    }
+
+    @Override
+    public String toString() {
+      return "Builder(prefix=" + prefix + ")";
+    }
+  }
 }